+++ Runtime/CesiumCreditSystem.cs
@@ -267,14 +267,23 @@ namespace CesiumForUnity
             return _defaultCreditSystem;
         }
 
-        internal bool HasLoadingImages()
+        /// <summary>
+        /// Gets the number of images on this credit system that are being loaded.
+        /// </summary>
+        /// <returns>The number of loading images.</returns>
+        internal int GetNumberOfLoadingImages()
         {
-            return this._numLoadingImages > 0;
+            return this._numLoadingImages;
         }
 
+        /// <summary>
+        /// A function invoked with StartCoroutine() to asynchronously load images from an HTML credit.
+        /// </summary>
+        /// <param name="url">A string containing either the base64-encoded image data or the URL of the image.</param>
         internal IEnumerator LoadImage(string url)
         {
             int index = this._images.Count;
+            this._numLoadingImages++;
 
             // Initialize a texture of arbitrary size as a placeholder,
             // so that when other images are loaded, their IDs align properly
@@ -287,23 +296,29 @@ namespace CesiumForUnity
                 // Load an image from a string that contains the
                 // "data:image/png;base64," prefix
                 string byteString = url.Substring(base64Prefix.Length);
-                byte[] bytes = Convert.FromBase64String(byteString);
-                if (!texture.LoadImage(bytes))
+                try
                 {
-                    Debug.Log("Could not parse image from base64 string.");
+                    byte[] bytes = Convert.FromBase64String(byteString);
+                    if (!texture.LoadImage(bytes))
+                    {
+                        Debug.Log("Credit image could not be loaded into Texture2D.");
+                    }
+                }
+                catch (FormatException e)
+                {
+                    Debug.Log("Could not parse credit image from base64 string.");
                 }
             }
             else
             {
                 // Load an image from a URL.
                 UnityWebRequest request = UnityWebRequestTexture.GetTexture(url);
-                this._numLoadingImages++;
                 yield return request.SendWebRequest();
 
                 if (request.result == UnityWebRequest.Result.ConnectionError ||
                     request.result == UnityWebRequest.Result.ProtocolError)
                 {
-                    Debug.Log(request.error);
+                    Debug.LogError(request.error);
                 }
                 else
                 {
@@ -314,10 +329,10 @@ namespace CesiumForUnity
                     UnityLifetime.Destroy(placeholderTexture);
                 }
 
-                this._numLoadingImages--;
             }
 
             texture.wrapMode = TextureWrapMode.Clamp;
+            this._numLoadingImages--;
         }
     }
 }